requirements.txt
future==0.18.3
iso8601==2.0.0
pip==23.1.2
pyserial==3.5
PyYAML==6.0
setuptools==67.8.0
wheel==0.40.0
